Remove redundant type information for bitfield elements, other than the signed/unsignedness of the element.
Change type of USB_SelectedPipe and USB_SelectedEndpoint for the AVR32 UC3 architecture to uint32_t to reduce the compiled code size.

+++ LUFA/Drivers/USB/Core/UC3/Endpoint_UC3.c
@@ -39,7 +39,7 @@
 uint8_t USB_ControlEndpointSize = ENDPOINT_CONTROLEP_DEFAULT_SIZE;
 #endif
 
-volatile uint8_t  USB_SelectedEndpoint = ENDPOINT_CONTROLEP;
+volatile uint32_t USB_SelectedEndpoint = ENDPOINT_CONTROLEP;
 volatile uint8_t* USB_EndpointFIFOPos[ENDPOINT_TOTAL_ENDPOINTS];
 
 bool Endpoint_ConfigureEndpoint_Prv(const uint8_t Number,
